// frontend/src/api/model/americanOptionRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Option Pricing Lab API
 * HTTP API for the FITE7405 mini option pricer.
 * OpenAPI spec version: 0.1.0
 */
import type { OptionType } from './optionType';

export interface AmericanOptionRequest {
  /** @exclusiveMinimum 0 */
  spot: number;
  /** @exclusiveMinimum 0 */
  strike: number;
  /** @exclusiveMinimum 0 */
  maturity: number;
  rate: number;
  /** @minimum 0 */
  volatility: number;
  /** @exclusiveMinimum 0 */
  steps: number;
  dividend_yield?: number;
  option_type: OptionType;
}