// frontend/src/api/model/arithmeticBasketOptionRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Option Pricing Lab API
 * HTTP API for the FITE7405 mini option pricer.
 * OpenAPI spec version: 0.1.0
 */
import type { OptionType } from './optionType';

export interface ArithmeticBasketOptionRequest {
  /** @exclusiveMinimum 0 */
  spot_1: number;
  /** @exclusiveMinimum 0 */
  spot_2: number;
  /** @exclusiveMinimum 0 */
  strike: number;
  /** @exclusiveMinimum 0 */
  maturity: number;
  rate: number;
  /** @minimum 0 */
  volatility_1: number;
  /** @minimum 0 */
  volatility_2: number;
  /**
     * @minimum -1
     * @maximum 1
     */
  correlation: number;
  option_type: OptionType;
  /** @exclusiveMinimum 1 */
  paths?: number;
  seed?: number;
  use_control_variate?: boolean;
}