// frontend/src/api/model/impliedVolatilityRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Option Pricing Lab API
 * HTTP API for the FITE7405 mini option pricer.
 * OpenAPI spec version: 0.1.0
 */
import type { OptionType } from './optionType';

export interface ImpliedVolatilityRequest {
  /** @exclusiveMinimum 0 */
  premium: number;
  /** @exclusiveMinimum 0 */
  spot: number;
  /** @exclusiveMinimum 0 */
  strike: number;
  /** @exclusiveMinimum 0 */
  maturity: number;
  rate: number;
  dividend_yield?: number;
  option_type: OptionType;
  /** @exclusiveMinimum 0 */
  initial_volatility?: number;
  /** @exclusiveMinimum 0 */
  tolerance?: number;
  /** @exclusiveMinimum 0 */
  max_iterations?: number;
  /** @exclusiveMinimum 0 */
  volatility_lower_bound?: number;
  /** @exclusiveMinimum 0 */
  volatility_upper_bound?: number;
}