// frontend/src/api/model/kikoPutOptionRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Option Pricing Lab API
 * HTTP API for the FITE7405 mini option pricer.
 * OpenAPI spec version: 0.1.0
 */

export interface KikoPutOptionRequest {
  /** @exclusiveMinimum 0 */
  spot: number;
  /** @exclusiveMinimum 0 */
  strike: number;
  /** @exclusiveMinimum 0 */
  maturity: number;
  rate: number;
  /** @minimum 0 */
  volatility: number;
  /** @minimum 0 */
  lower_barrier: number;
  /** @exclusiveMinimum 0 */
  upper_barrier: number;
  /** @exclusiveMinimum 0 */
  observations: number;
  /** @minimum 0 */
  rebate: number;
  /** @exclusiveMinimum 1 */
  paths?: number;
  /** @exclusiveMinimum 0 */
  delta_shift?: number;
}